+++ CHANGES.rst
@@ -43,6 +43,9 @@ Changelog
 - Fixed bug: moderation_form fail to render when using VirtualHostMonster
   [rafaelbco]
 
+- Added check in ploneboard tool that allows anonymous users to insert a name in comments/discussions
+  [cekk]
+
 3.4 (2013-05-11)
 ----------------